refactor: extract WorkspaceState media into an extension file
Move attachments, downloads, voice recording/metering, the per-conversation
media-download state stores, and composer-draft clearing into
WorkspaceState+Media.swift. Pure code movement, no behavior change.
Build green; all 208 tests pass.
