+++ whitenoise-mac/Core/WorkspaceState.swift
@@ -374,33 +374,6 @@ final class WorkspaceState {
         return ComposerDraftKey(accountId: activeAccountId, chatId: chatId)
     }
 
-    func clearAllComposerDrafts() {
-        draftTextByConversation.removeAll()
-        replyDraftContextByConversation.removeAll()
-        pendingMediaAttachmentsByConversation.removeAll()
-    }
-
-    func clearComposerDrafts(for chatIds: [String], accountId: String) {
-        for chatId in chatIds {
-            let key = ComposerDraftKey(accountId: accountId, chatId: chatId)
-            draftTextByConversation[key] = nil
-            replyDraftContextByConversation[key] = nil
-            pendingMediaAttachmentsByConversation[key] = nil
-        }
-    }
-
-    func clearComposerDrafts(forAccountId accountId: String) {
-        for key in draftTextByConversation.keys.filter({ $0.accountId == accountId }) {
-            draftTextByConversation[key] = nil
-        }
-        for key in replyDraftContextByConversation.keys.filter({ $0.accountId == accountId }) {
-            replyDraftContextByConversation[key] = nil
-        }
-        for key in pendingMediaAttachmentsByConversation.keys.filter({ $0.accountId == accountId }) {
-            pendingMediaAttachmentsByConversation[key] = nil
-        }
-    }
-
     let clientFactory: @MainActor () throws -> any MarmotRuntime
     let localNotificationCenter: any LocalNotificationCenter
     let appActivityProvider: @MainActor () -> Bool
@@ -1246,78 +1219,6 @@ final class WorkspaceState {
         }
     }
 
-    func mediaDownloadState(for message: MessageItem, attachment: MessageMediaAttachment) -> MediaDownloadState {
-        mediaDownloadStateStore(for: message, attachment: attachment).state
-    }
-
-    func mediaDownloadStateStore(
-        for message: MessageItem,
-        attachment: MessageMediaAttachment
-    ) -> MediaDownloadStateStore {
-        mediaDownloadStateStore(forKey: mediaDownloadKey(message: message, attachment: attachment))
-    }
-
-    func loadMediaAttachment(_ attachment: MessageMediaAttachment, for message: MessageItem) async {
-        let key = mediaDownloadKey(message: message, attachment: attachment)
-        let stateStore = mediaDownloadStateStore(forKey: key)
-        if case .loaded = stateStore.state {
-            return
-        }
-        if case .loading = stateStore.state {
-            return
-        }
-
-        guard let client, let activeAccount, !message.groupIdHex.isEmpty else {
-            stateStore.update(.failed(L10n.string("Attachment unavailable")))
-            return
-        }
-
-        let accountId = activeAccount.id
-        let accountRef = activeAccount.accountRef
-        let groupIdHex = message.groupIdHex
-        stateStore.update(.loading)
-
-        do {
-            let reference = try await resolvedMediaReference(
-                attachment.reference,
-                accountRef: accountRef,
-                groupIdHex: groupIdHex,
-                client: client
-            )
-            let download = try await client.downloadMedia(
-                accountRef: accountRef,
-                groupIdHex: groupIdHex,
-                reference: reference
-            )
-            guard activeAccountId == accountId else { return }
-            stateStore.update(
-                .loaded(
-                    MessageMediaDownload(
-                        data: download.plaintext,
-                        fileName: download.fileName,
-                        mediaType: download.mediaType,
-                        sizeBytes: download.sizeBytes
-                    )
-                )
-            )
-        } catch {
-            guard activeAccountId == accountId else { return }
-            stateStore.update(.failed(error.localizedDescription))
-        }
-    }
-
-    /// Lazily allocates per-attachment stores from SwiftUI body lookup without observing the
-    /// backing dictionary; `mediaDownloads` is `@ObservationIgnored`, and pruning bounds it to
-    /// the active conversation.
-    func mediaDownloadStateStore(forKey key: String) -> MediaDownloadStateStore {
-        if let store = mediaDownloads[key] {
-            return store
-        }
-        let store = MediaDownloadStateStore()
-        mediaDownloads[key] = store
-        return store
-    }
-
     /// The bech32 `npub` form of a hex public key — the canonical, user-facing way to show
     /// a Nostr public key. Falls back to the hex until the account cache has been hydrated
     /// off the main thread.
@@ -1378,221 +1279,12 @@ final class WorkspaceState {
         )
     }
 
-    func addMediaAttachments(from urls: [URL]) async {
-        guard let draftKey = selectedComposerDraftKey else { return }
-        guard canBeginMediaAttachmentSelection() else { return }
-        let fileURLs = urls.filter(\.isFileURL)
-        guard !fileURLs.isEmpty else { return }
-
-        let selected = Array(fileURLs.prefix(remainingMediaAttachmentSlots))
-        if selected.count < fileURLs.count {
-            presentMaxMediaAttachmentWarning()
-        }
-
-        for url in selected {
-            let isSecurityScoped = url.startAccessingSecurityScopedResource()
-            defer {
-                if isSecurityScoped {
-                    url.stopAccessingSecurityScopedResource()
-                }
-            }
-            do {
-                let attachment = try await OutgoingMediaDraftProcessor.preparedAttachment(fromFileURL: url)
-                appendPendingMediaAttachment(attachment, for: draftKey)
-            } catch is CancellationError {
-                return
-            } catch {
-                lastError = error.localizedDescription
-            }
-        }
-    }
-
-    func removePendingMediaAttachment(_ id: PendingMediaAttachment.ID) {
-        guard let selectedComposerDraftKey else { return }
-        var attachments = pendingMediaAttachmentsByConversation[selectedComposerDraftKey] ?? []
-        attachments.removeAll { $0.id == id }
-        pendingMediaAttachmentsByConversation[selectedComposerDraftKey] = attachments.isEmpty ? nil : attachments
-    }
-
-    func toggleVoiceRecording() async {
-        if isRecordingVoiceMessage {
-            await finishVoiceRecording()
-        } else {
-            await startVoiceRecording()
-        }
-    }
-
-    func startVoiceRecording() async {
-        guard !isRecordingVoiceMessage else { return }
-        guard canBeginMediaAttachmentSelection() else { return }
-
-        let hasPermission = await requestMicrophoneAccess()
-        guard hasPermission else {
-            lastError = L10n.string("Microphone access is needed to record voice messages.")
-            return
-        }
-
-        let directory = FileManager.default.temporaryDirectory
-            .appendingPathComponent("WhiteNoiseVoiceRecordings", isDirectory: true)
-        do {
-            try FileManager.default.createDirectory(at: directory, withIntermediateDirectories: true)
-            let fileName = "voice-\(Int(Date().timeIntervalSince1970)).m4a"
-            let url = directory.appendingPathComponent(fileName)
-            let settings: [String: Any] = [
-                AVFormatIDKey: Int(kAudioFormatMPEG4AAC),
-                AVSampleRateKey: 44_100,
-                AVNumberOfChannelsKey: 1,
-                AVEncoderAudioQualityKey: AVAudioQuality.medium.rawValue,
-            ]
-            let recorder = try AVAudioRecorder(url: url, settings: settings)
-            recorder.isMeteringEnabled = true
-            guard recorder.record() else {
-                throw VoiceRecordingFailure.startFailed
-            }
-
-            voiceRecorder = recorder
-            voiceRecordingURL = url
-            voiceRecordingSamples = []
-            voiceRecordingDurationSeconds = 0
-            isRecordingVoiceMessage = true
-            startVoiceRecordingMetering()
-        } catch {
-            resetVoiceRecording(deleteFile: true)
-            lastError = L10n.string("Voice recording could not start.")
-        }
-    }
-
-    func finishVoiceRecording() async {
-        guard isRecordingVoiceMessage, let recorder = voiceRecorder, let url = voiceRecordingURL else {
-            resetVoiceRecording(deleteFile: true)
-            return
-        }
-        let draftKey = selectedComposerDraftKey
-        let duration = max(voiceRecordingDurationSeconds, recorder.currentTime)
-        let samples = voiceRecordingSamples
-        let fileName = url.lastPathComponent
-        recorder.stop()
-        resetVoiceRecording(deleteFile: false)
-
-        guard let draftKey else {
-            try? FileManager.default.removeItem(at: url)
-            return
-        }
-
-        do {
-            let attachment = try await OutgoingMediaDraftProcessor.preparedVoiceAttachment(
-                from: VoiceRecordingResult(
-                    url: url,
-                    fileName: fileName,
-                    durationSeconds: duration,
-                    waveformSamples: samples
-                )
-            )
-            appendPendingMediaAttachment(attachment, for: draftKey)
-        } catch is CancellationError {
-            return
-        } catch {
-            lastError = error.localizedDescription
-        }
-    }
-
-    func cancelVoiceRecording() {
-        resetVoiceRecording(deleteFile: true)
-    }
-
     var remainingMediaAttachmentSlots: Int {
         max(0, OutgoingMediaDraftProcessor.maxAttachmentCount - pendingMediaAttachments.count)
     }
 
-    func canBeginMediaAttachmentSelection() -> Bool {
-        guard client != nil, selectedChat != nil else { return false }
-        guard remainingMediaAttachmentSlots > 0 else {
-            presentMaxMediaAttachmentWarning()
-            return false
-        }
-        return true
-    }
-
-    func appendPendingMediaAttachment(_ attachment: PendingMediaAttachment, for draftKey: ComposerDraftKey) {
-        var attachments = pendingMediaAttachmentsByConversation[draftKey] ?? []
-        if attachment.kind == .audio {
-            attachments.removeAll { $0.kind == .audio }
-        }
-        guard attachments.count < OutgoingMediaDraftProcessor.maxAttachmentCount else {
-            presentMaxMediaAttachmentWarning()
-            return
-        }
-        attachments.append(attachment)
-        pendingMediaAttachmentsByConversation[draftKey] = attachments
-        if attachment.kind == .audio {
-            draftTextByConversation[draftKey] = nil
-        }
-    }
-
-    func presentMaxMediaAttachmentWarning() {
-        lastError = String(
-            format: L10n.string("You can send up to %lld attachments at once"),
-            Int64(OutgoingMediaDraftProcessor.maxAttachmentCount)
-        )
-    }
-
     enum VoiceRecordingFailure: Error {
         case startFailed
-    }
-
-    func requestMicrophoneAccess() async -> Bool {
-        switch AVCaptureDevice.authorizationStatus(for: .audio) {
-        case .authorized:
-            return true
-        case .notDetermined:
-            return await withCheckedContinuation { continuation in
-                AVCaptureDevice.requestAccess(for: .audio) { granted in
-                    continuation.resume(returning: granted)
-                }
-            }
-        case .denied, .restricted:
-            return false
-        @unknown default:
-            return false
-        }
-    }
-
-    func startVoiceRecordingMetering() {
-        voiceRecordingMeterTask?.cancel()
-        voiceRecordingMeterTask = Task { @MainActor [weak self] in
-            while !Task.isCancelled {
-                do {
-                    try await Task.sleep(nanoseconds: 70_000_000)
-                } catch {
-                    return
-                }
-                guard let self, let recorder = self.voiceRecorder else { return }
-                recorder.updateMeters()
-                self.voiceRecordingDurationSeconds = recorder.currentTime
-                let power = recorder.averagePower(forChannel: 0)
-                let normalized = max(0.05, min(1, CGFloat(pow(10, power / 36))))
-                self.voiceRecordingSamples.append(normalized)
-                if self.voiceRecordingSamples.count > MediaWaveformAnalyzer.sampleCount {
-                    self.voiceRecordingSamples.removeFirst(
-                        self.voiceRecordingSamples.count - MediaWaveformAnalyzer.sampleCount)
-                }
-            }
-        }
-    }
-
-    func resetVoiceRecording(deleteFile: Bool) {
-        voiceRecordingMeterTask?.cancel()
-        voiceRecordingMeterTask = nil
-        voiceRecorder?.stop()
-        voiceRecorder = nil
-        let url = voiceRecordingURL
-        voiceRecordingURL = nil
-        isRecordingVoiceMessage = false
-        voiceRecordingSamples = []
-        voiceRecordingDurationSeconds = 0
-        if deleteFile, let url {
-            try? FileManager.default.removeItem(at: url)
-        }
     }
 
     func restoreOrSelectFirstAccount() {
@@ -1819,52 +1511,6 @@ final class WorkspaceState {
         )
         privacySecuritySettings.telemetryCredentialsAvailable = config.telemetryCredentialsAvailable
         privacySecuritySettings.auditLogCredentialsAvailable = config.auditLogCredentialsAvailable
-    }
-
-    func mediaDownloadKey(message: MessageItem, attachment: MessageMediaAttachment) -> String {
-        [activeAccountId ?? "", message.groupIdHex, attachment.id].joined(separator: "\u{1F}")
-    }
-
-    func pruneMediaDownloadCache(keeping groupIdHex: String?) {
-        guard let activeAccountId, let groupIdHex else {
-            resetMediaDownloadStateStores()
-            return
-        }
-
-        let prefix = [activeAccountId, groupIdHex, ""].joined(separator: "\u{1F}")
-        let removedKeys = mediaDownloads.keys.filter { !$0.hasPrefix(prefix) }
-        for key in removedKeys {
-            // Notify any lingering per-attachment observers before dropping the store.
-            mediaDownloads[key]?.update(.idle)
-            mediaDownloads[key] = nil
-        }
-    }
-
-    func resetMediaDownloadStateStores() {
-        for store in mediaDownloads.values {
-            // Notify any lingering per-attachment observers before clearing the cache.
-            store.update(.idle)
-        }
-        mediaDownloads.removeAll()
-    }
-
-    func resolvedMediaReference(
-        _ reference: MediaAttachmentReferenceFfi,
-        accountRef: String,
-        groupIdHex: String,
-        client: any MarmotRuntime
-    ) async throws -> MediaAttachmentReferenceFfi {
-        guard reference.sourceEpoch == 0 else {
-            return reference
-        }
-
-        let records = try await runOffMain {
-            try client.listMedia(accountRef: accountRef, groupIdHex: groupIdHex, limit: nil)
-        }
-        return records.first { record in
-            record.reference.plaintextSha256 == reference.plaintextSha256
-                || record.reference.ciphertextSha256 == reference.ciphertextSha256
-        }?.reference ?? reference
     }
 
     func resetNewChatComposer() {
